+++ ParseSwift.playground/Sources/Common.swift
@@ -5,7 +5,8 @@ public func initializeParse() {
     ParseSwift.initialize(applicationId: "applicationId",
                      clientKey: "clientKey",
                      masterKey: "masterKey",
-                     serverURL: URL(string: "http://localhost:1337/1")!)
+                     serverURL: URL(string: "http://localhost:1337/1")!,
+                     useTransactionsInternally: false)
 }
 
 public func initializeParseCustomObjectId() {