+++ CHANGELOG.md
@@ -1,11 +1,17 @@
 # Parse-Swift Changelog
 
 ### main
-[Full Changelog](https://github.com/parse-community/Parse-Swift/compare/1.8.0...main)
+[Full Changelog](https://github.com/parse-community/Parse-Swift/compare/1.8.1...main)
 * _Contributing to this repo? Add info about your change here to be included in the next release_
+
+### 1.8.0
+[Full Changelog](https://github.com/parse-community/Parse-Swift/compare/1.8.0...1.8.1)
 
 __Improvements__
 - Append instead of replace when using query select, exclude, include, and fields ([#155](https://github.com/parse-community/Parse-Swift/pull/155)), thanks to [Corey Baker](https://github.com/cbaker6).
+
+__Fixes__
+- Transactions currently don't work when using mongoDB(postgres does work) on the parse-server. Internal use of transactions are disabled by default. If you want the Swift SDK to use transactions internally, you need to set useTransactionsInternally=true when configuring the client. It is recommended not to use transactions if you are using mongoDB until it's fixed on the server ([#158](https://github.com/parse-community/Parse-Swift/pull/158)), thanks to [Corey Baker](https://github.com/cbaker6).
 
 ### 1.8.0
 [Full Changelog](https://github.com/parse-community/Parse-Swift/compare/1.7.2...1.8.0)